third_party/blink/web_tests/virtual/customizable-select-disabled/html5lib/generated/run-webkit02-data-expected.txt

CONSOLE WARNING: A xh<optgroup tag was parsed inside of a <select> which was not inserted into the document. This is not valid HTML and the behavior may be changed in future versions of chrome.
../resources/webkit02.dat:
12

Test 12 of 16 in ../resources/webkit02.dat failed. Input:
<isindex action="x">
Got:
| <isindex>
|   action="x"
Expected:
| <form>
|   action="x"
| <hr>
| <label>
|   "This is a searchable index. Enter search keywords: "
|   <input>
|     name="isindex"
| <hr>
